libavcodec update (most of it is stuff like INT32 -> int32_t). de-tab-ify thor's mods to existing files =)
git-svn-id: http://svn.mythtv.org/svn/trunk@972 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libavcodec/audresample.c
@@ -20,8 +20,8 @@
 
 typedef struct {
     /* fractional resampling */
-    UINT32 incr; /* fractional increment */
-    UINT32 frac;
+    uint32_t incr; /* fractional increment */
+    uint32_t frac;
     int last_sample;
     /* integer down sample */
     int iratio;  /* integer divison ratio */
@@ -43,7 +43,7 @@ struct ReSampleContext {
 static void init_mono_resample(ReSampleChannelContext *s, float ratio)
 {
     ratio = 1.0 / ratio;
-    s->iratio = (int)floor(ratio);
+    s->iratio = (int)floorf(ratio);
     if (s->iratio == 0)
         s->iratio = 1;
     s->incr = (int)((ratio / s->iratio) * FRAC);
